Handle deleting read-only files on Windows
Should fix #74.
